refactor: move the amd64 encoder under src/core/encoder
