+++ src/core/encoder/amd64/asm.go
