+++ src/core/encoder/amd64/asm_sse.go
