+++ src/core/encoder/amd64/module.go
